move blockexpr tests to test_linearize
